As discussed in https://github.com/dotnet/coreclr/issues/10642#issuecomment-321104523, I've done some perf work on core decimal functions. The goal is to get faster than the native (OleAut) implementation currently used in CoreCLR and then move System.Decimal to the shared partition.

Haven't done any 32-bit specific optimizations, so that might need some more work. @jkotas is this approach OK? I added CMP first because I changed it the most. I've been testing it by binary-comparing the results between native/CoreRT/CoreRT2, but I imagine this would be a bit overkill for CoreCLR tests (large parts of Decimal[.DecCalc].cs would have to be included in the tests folder)? For the CoreRT build I used pre-recorded input and expected output files, but for 440 different decimals it's ~9MB. Would it make sense to trim the input set down to for example 100 decimals and include the ~0.5MB reference data file in tests?

---

Perf measured using a weighted [random distribution of decimals](https://gist.github.com/pentp/260d9641ce4490c94f708d64e7fbd7c8) that might better represent real world use (gist is sorted for better overview). A total of 440^2=~190K combinations were tested (excluding overflows/div-by-0).

``` ini

BenchmarkDotNet=v0.10.9, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), ProcessorCount=8
.NET Core SDK=
  [Host] : .NET Core ? (Framework 4.6.25731.0), 64bit RyuJIT

EvaluateOverhead=False  Toolchain=InProcessToolchain  WarmupCount=1  
```
### 64-bit
 |  CMP |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 21.83 ns | 0.0176 ns | 0.0165 ns |   1.00 |
 |  CoreRT | 38.32 ns | 0.0465 ns | 0.0435 ns |   1.75 |
 | CoreRT2 | 11.13 ns | 0.0038 ns | 0.0030 ns |   0.50 |

 |  ADD |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 28.05 ns | 0.0121 ns | 0.0108 ns |   1.00 |
 |  CoreRT | 50.58 ns | 0.0944 ns | 0.0883 ns |   1.80 |
 | CoreRT2 | 23.99 ns | 0.0184 ns | 0.0172 ns |   0.86 |

 |  MUL |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 29.86 ns | 0.0063 ns | 0.0056 ns |   1.00 |
 |  CoreRT | 31.38 ns | 0.0348 ns | 0.0326 ns |   1.05 |
 | CoreRT2 | 18.67 ns | 0.0160 ns | 0.0150 ns |   0.62 |

 |  DIV |      Mean |     Error |    StdDev | Scaled |
 |-------- |----------:|----------:|----------:|-------:|
 |  Native | 116.90 ns | 0.3287 ns | 0.2745 ns |   1.00 |
 |  CoreRT | 183.50 ns | 0.2975 ns | 0.2783 ns |   1.57 |
 | CoreRT2 |  78.49 ns | 0.0675 ns | 0.0631 ns |   0.67 |

### 32-bit
 |  CMP |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 16.53 ns | 0.0087 ns | 0.0078 ns |   1.00 |
 |  CoreRT | 48.12 ns | 0.0387 ns | 0.0362 ns |   2.91 |
 | CoreRT2 | 15.90 ns | 0.0160 ns | 0.0142 ns |   0.96 |

 | ADD |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 28.84 ns | 0.0097 ns | 0.0091 ns |   1.00 |
 |  CoreRT | 65.52 ns | 0.0695 ns | 0.0650 ns |   2.27 |
 | CoreRT2 | 36.35 ns | 0.0075 ns | 0.0063 ns |   1.26 |

 |  MUL |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 19.26 ns | 0.0144 ns | 0.0134 ns |   1.00 |
 |  CoreRT | 49.18 ns | 0.0704 ns | 0.0658 ns |   2.55 |
 | CoreRT2 | 28.61 ns | 0.0186 ns | 0.0155 ns |   1.48 |

 |  DIV |     Mean |     Error |    StdDev | Scaled |
 |-------- |---------:|----------:|----------:|-------:|
 |  Native | 121.2 ns | 0.0571 ns | 0.0476 ns |   1.00 |
 |  CoreRT | 236.1 ns | 0.1411 ns | 0.1320 ns |   1.95 |
 | CoreRT2 | 122.5 ns | 0.0959 ns | 0.0897 ns |   1.01 |

### 64-bit CoreRT (same test data, basic measurement code):

 |  Method |     Before |     After |
 |-------- |---------:|---------:|
 |  CMP | 34.1 ns | 12.4 ns |
 |  ADD | 50.0 ns | 25.2 ns |
 |  MUL | 31.9 ns | 19.0 ns |
 |  DIV | 183.3 ns | 79.6 ns |