@jkotas 2017-09-05
>  440 different decimals it's ~9MB

Right, 9MB file is overkill.

Ideally, the Decimal tests for this would live in CoreFX repo. For the extensive testing that tries many different combinations, would it make sense to generate random numbers, compute the result using BigInteger and verify that decimal got the same answer?