@pentp 2017-09-05
> compute the result using BigInteger and verify that decimal got the same answer?

It could cover some of the calculations, but getting the proper scaling, rounding and corner cases (overflows) would be more difficult. But I'll look into it.