@pentp 2017-09-08
I don't know what's up with the two pending builds, but this should be ready for merging now.
To get decimal ready for moving to the shared partition, I plan to get 32-bit perf closer to parity with native and go over some of the remaining common functions also (casting, rounding, formatting and parsing support).

I've been thinking how to add comprehensive tests to CoreFX and the fixed reference data approach with a relative small dataset looks easier and allows for exact binary compatibility with existing code. I haven't yet tried the `BigInteger` approach, so I might be somewhat biased here.