@jkotas 2017-09-10
@pentp Thank you! Both the code and the perf improvement look pretty good.

I think we need to look at the test strategy before merging this. Let's wait until you get the test PR for CoreFX.

cc @joperezr @AlexGhiondea as System.Runtime owners