@jkotas 2017-09-11
I have done some ad hoc testing on this and I found bugs. For example:

```
    static void Main()
    {
        var a = new Decimal(123877878, -16789245, 1086421879, true, 16);
        var b = new Decimal(527635459, -80701438, 1767087216, true, 24);
        var res = new Decimal(253511426, -909347550, -753557281, false, 12);
        Console.WriteLine((a * b == res) ? "PASSED" : "FAILED");
    }
```

Passes before this change, fails with this change.