@pentp 2017-09-12
I've got ADD + CMP [tests based on BigInteger,](https://gist.github.com/pentp/706230432380b1cbedde9f7d50cc7e4e) but getting the scaling and rounding right was painful. About 3M combinations are tested in ~2 seconds at the moment.

CoreFX tests are a bit differently structured than CoreCLR tests and using Xunit. Is it OK to structure this as a single static data initialization method + a [Fact] method for each of ADD/MUL/DIV/CMP?