@jkotas 2017-09-12
I do not have a strong opinions on it, also I think that the answer depends on the size of the static data. You can start a WIP PR in corefx with what you have and ask for the feedback there.