@pentp 2017-09-25
So the tests are in CoreFX now. I'm not sure how to run them on CoreRT, but I've run basically the same test code locally on CoreRT.