@jkotas 2017-09-26
@pentp Could you please resolve the merge conflict? Otherwise, I think this is good to go. Thank you!