@pentp 2017-09-26
I rebased and resolved the merge conflict.