@tannergooding 2017-09-26
@jkotas, @pentp. Is there a plan to move `System.Decimal` to be shared between CoreCLR and CoreRT?