@pentp 2017-09-26
32-bit perf isn't quite there yet and some other commonly used functions need measuring, improvements and new tests. @jkotas knows better what the actual steps/plan is, but I started decimal related work because I hope to use it in my decimal-heavy web application one day and learn new things along the way.