@jkotas 2017-09-26
As @pentp said - once the perf is better or close enough, we can move it to shared partition.