@MichalStrehovsky 2017-10-03
@pentp @jkotas We haven't had a successful rolling test run since this change got merged; see latest logs here: https://ci.dot.net/job/dotnet_corert/job/master/job/debug_windows_nt/1216/#showFailuresLink

The tests that are failing are e.g. this one: https://github.com/dotnet/coreclr/blob/master/tests/src/JIT/Methodical/fp/exgen/1000w1d.cs

We're seeing unexpected System.OverflowException and there are other failures too.