@jkotas 2017-10-03
I am going to revert the change until we get it sorted out. @pentp Could you please take a look at the failures? (https://github.com/dotnet/corert/blob/master/Documentation/how-to-run-tests.md has instructions how to run the CoreCLR tests.)