@trylek 2017-10-03
@pentp I found one particular typo in the change - in Decimal.cs at line 180, in the constructor accepting long, there should be

            Low64 = (ulong)value_copy;

instead of

            Low64 = (ulong)value;

otherwise it doesn't work for negative values.

Thanks

Tomas