@pentp 2017-10-03
I can't get the CoreCLR tests working on CoreRT, first it called the wrong `vcvarsall.bat`, then it had trouble downloading the tests.zip file, after fixing these I'm stuck at `Visual Studio 2013+ (Community is free) is a prerequisite to build this repository` (I'm running this in "x64 Native Tools Command Prompt for VS 2017").
`tests\runtest.cmd x64 release /mode ryujit` works, `tests\runtest.cmd x64 release /mode ryujit /coreclr Top200` doesn't.

But I did run this specific test (1000w1d.cs) and after fixing the `Low64 = (ulong)value_copy` bug it works - I had somehow managed to mangle this line when copying it from the performance test code where I had the correct code...

So how should I make the fix - restore the branch for this PR and add a commit to it?