@jkotas 2017-10-04
>  I'm stuck at `Visual Studio 2013+ (Community is free)` is a prerequisite to build this repository

I have cleaned up the VS auto-detection for CoreCLR tests. It should work now - always against VS2017.

> restore the branch for this PR and add a commit to it?

Yes. Either restore the branch, or revert the revert.

Thanks!